+++ test/BundleBuilderTest.ts
@@ -29,7 +29,14 @@ suite("BundleBuilder", function MemoryStoreTest() {
             fs: require.resolve("fs"),
         };
 
-        var bundleBldr = BundleBuilder.buildBundler((opts: TsBrowserify.Options) => new TsBrowserify(opts), /*watchify*/null, {
+        var bundlePaths: CodePaths = {
+            entryFile: "./test/test-proj/App.js",
+            dstDir: "./test/tmp/",
+            srcPaths: ["./test-proj", "node_modules"],
+            projectRoot: process.cwd() + "/test"
+        };
+
+        var bundleOpts = BundleBuilder.createOptions<TsBrowserify.Options>({
             debug: true,
             rebuild: false,
             verbose: true,
@@ -37,8 +44,11 @@ suite("BundleBuilder", function MemoryStoreTest() {
             depsSort,
             moduleDeps,
             insertModuleGlobals,
-        }, BundleBuilder.compileBundle)
-        .setBundleListeners({
+        }, bundlePaths, null);
+
+        var bundler = new TsBrowserify(bundleOpts);
+
+        BundleBuilder.compileBundle(bundler, bundleOpts, bundlePaths.dstDir, (br) => BundleBuilder.createDefaultBundler(br, { dstFileName: "bundle.js" }), [], {
             finishAll: () => {
                 var bundleMap: any;
 
@@ -65,14 +75,6 @@ suite("BundleBuilder", function MemoryStoreTest() {
 
                 done();
             },
-        })
-        .compileBundle({
-            entryFile: "./test/test-proj/App.js",
-            dstDir: "./test/tmp/",
-            srcPaths: ["./test-proj", "node_modules"],
-            projectRoot: process.cwd() + "/test"
-        }, {
-            dstFileName: "bundle.js"
         });
     });
 
@@ -91,11 +93,11 @@ suite("BundleBuilder", function MemoryStoreTest() {
                 bundles: [{
                     dstFileName: "bundle.js",
                     dstMapFile: "bundle.js.map",
-                    prelude: "// main bundle prelude test\n" + bOpts.prelude,
+                    prelude: "// main bundle prelude test\n" + bundleOpts.prelude,
                 }, {
                     dstFileName: "bundle-data-access.js",
                     dstMapFile: "bundle-data-access.js.map",
-                    prelude: "// data access bundle prelude test\n" + bOpts.typescriptHelpers + "var require = " + bOpts.prelude,
+                    prelude: "// data access bundle prelude test\n" + bundleOpts.typescriptHelpers + "var require = " + bundleOpts.prelude,
                     preludePath: "./_prelude-with-typescript-helpers.js",
                 }],
                 destinationPicker: function (path) {
@@ -104,19 +106,21 @@ suite("BundleBuilder", function MemoryStoreTest() {
             });
         });
 
-        var allDeps: { [name: string]: string[] };
-        var bOpts: TsBrowserify.Options;
-        var savedDeps = [];
-        var bundleBldr = BundleBuilder.buildBundler<TsBrowserify, TsBrowserify.Options>((opts) => {
-            var bundler = new TsBrowserify(bOpts = opts);
-            allDeps = BrowserifyHelper.addDependencyTracker(process.cwd() + "\\test\\", bundler).allDeps;
-            return bundler;
-        }, null/*(b, opts) => TsWatchify(b, { delay: 500 })*/, {
+        var bundlePaths = {
+            entryFile: "./test/test-proj/App.js",
+            dstDir: "./test/tmp/",
+            srcPaths: ["./test-proj", "node_modules"],
+            projectRoot: process.cwd() + "/test"
+        };
+
+        var bundleOpts = BundleBuilder.createOptions<TsBrowserify.Options>({
             debug: true,
             rebuild: false,
             verbose: true,
             typescript: { includeHelpers: true },
-            browserPack: function (opts) { return packer.createPackStreams().baseStream; },
+            browserPack: function (opts) {
+                return packer.createPackStreams().baseStream;
+            },
             depsSort: depsSort,
             moduleDeps: function (opts) {
                 opts.detect = function (src) {
@@ -126,12 +130,13 @@ suite("BundleBuilder", function MemoryStoreTest() {
                 return moduleDeps(opts);
             },
             insertModuleGlobals: insertModuleGlobals,
-        }, BundleBuilder.compileBundle)
-        .setBundleSourceCreator(packer.multiBundleSourceCreator)
-        .transforms(function (browserify) {
-            return [];
-        })
-        .setBundleListeners({
+        }, bundlePaths, null/*(b, opts) => TsWatchify(b, { delay: 500 })*/);
+
+        var bundler = new TsBrowserify(bundleOpts);
+
+        var allDeps = BrowserifyHelper.addDependencyTracker(process.cwd() + "\\test\\", bundler).allDeps;
+
+        BundleBuilder.compileBundle(bundler, bundleOpts, bundlePaths.dstDir, packer.multiBundleSourceCreator, [], {
             finishAll: function () {
                 //console.log("deps", JSON.stringify(allDeps, undefined, "  "));
                 var bundleDataSourceMap: any;
@@ -146,12 +151,12 @@ suite("BundleBuilder", function MemoryStoreTest() {
                         asr.isTrue(FileUtil.existsFileSync("./test/tmp/bundle.js.map"));
                         asr.isTrue(FileUtil.existsFileSync("./test/tmp/bundle-data-access.js"));
                         asr.isTrue(FileUtil.existsFileSync("./test/tmp/bundle-data-access.js.map"));
-                        //fs.unlinkSync("./test/tmp/bundle.js");
-                        //fs.unlinkSync("./test/tmp/bundle.js.map");
-                        //fs.unlinkSync("./test/tmp/bundle-data-access.js");
-                        //fs.unlinkSync("./test/tmp/bundle-data-access.js.map");
-                        //fs.rmdirSync("./test/tmp");
-                        //asr.isNotTrue(FileUtil.existsDirSync("./test/tmp"));
+                        fs.unlinkSync("./test/tmp/bundle.js");
+                        fs.unlinkSync("./test/tmp/bundle.js.map");
+                        fs.unlinkSync("./test/tmp/bundle-data-access.js");
+                        fs.unlinkSync("./test/tmp/bundle-data-access.js.map");
+                        fs.rmdirSync("./test/tmp");
+                        asr.isNotTrue(FileUtil.existsDirSync("./test/tmp"));
                     }
                 });
 
@@ -175,13 +180,7 @@ suite("BundleBuilder", function MemoryStoreTest() {
 
                 done();
             }
-        })
-        .compileBundle({
-            entryFile: "./test/test-proj/App.js",
-            dstDir: "./test/tmp/",
-            srcPaths: ["./test-proj", "node_modules"],
-            projectRoot: process.cwd() + "/test"
-        }, null);
+        });
     });